feat: implement vendor management with CRUD operations and validation and fix requests (#5)

+++ app/services/Intro/IntroHelperFunctionService.php
@@ -20,7 +20,7 @@ class IntroHelperFunctionService
      */
     public function getIntroData($id = null)
     {
-        return $id ? Intro::findOrFail($id) : Intro::allActive();
+        return $id ? Intro::find($id) : Intro::allActive();
     }
 
     public function getIntro()
@@ -52,7 +52,10 @@ class IntroHelperFunctionService
     public function updateIntroData($request, $id)
     {
         return DB::transaction(function () use ($request, $id) {
-            $intro = Intro::findOrFail($id);
+            $intro = Intro::find($id);
+            if (!$intro) {
+                return false;
+            }
 
             if (!empty($request['image'])) {
                 $request['image'] = FileService::replace(
@@ -80,9 +83,10 @@ class IntroHelperFunctionService
     public function deleteIntro($id)
     {
         return DB::transaction(function () use ($id) {
-            $intro = Intro::findOrFail($id);
-
-            // Delete the intro record and its associated image
+            $intro = Intro::find($id);
+            if (!$intro) {
+                return false;
+            }
             FileService::delete($intro->getRawOriginal('image'));
             $intro->delete();
